fixed testing bugs

+++ test/helper.rb
@@ -13,7 +13,7 @@ class Test::Unit::TestCase
     generated_test_files.each do |x|
       if File.file?(x)
         FileUtils.rm(x)
-      elsif File.directory?(x) && x != "#{JumpStart::ROOT_PATH}/test/destination_dir"
+      elsif File.directory?(x) && x != "#{JumpStart::ROOT_PATH}/test/destination_dir" && Dir.exists?(x)
         FileUtils.remove_dir(x)
       end
     end
